refactor: move mock file (#76)
